requirement.txt
OpenAI
langchain
langchain-openai
langchain-community
langchain-core
streamlit
wikipedia
python-dotenv
